requirements_gui.txt
# GUI Voice Assistant - Requirements

# Core GUI
customtkinter>=5.2.0

# Voice Recognition
SpeechRecognition>=3.10.0
PyAudio>=0.2.13

# Text-to-Speech
pyttsx3>=2.90

# AI - Gemini (Stable)
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
